Move item-search-config into item-search-filter
