+++ src/app/search/item-search-filter.ts
@@ -1,3 +1,5 @@
+import { DestinyVersion } from '@destinyitemmanager/dim-api-types';
+import { destinyVersionSelector } from 'app/accounts/selectors';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { customStatsSelector, languageSelector } from 'app/dim-api/selectors';
 import { DimLanguage } from 'app/i18n';
@@ -5,6 +7,7 @@ import { TagValue } from 'app/inventory/dim-item-info';
 import { d2ManifestSelector } from 'app/manifest/selectors';
 import { Settings } from 'app/settings/initial-settings';
 import { WishListRoll } from 'app/wishlists/types';
+import memoizeOne from 'memoize-one';
 import { createSelector } from 'reselect';
 import { DimItem } from '../inventory/item-types';
 import {
@@ -22,12 +25,64 @@ import { querySelector } from '../shell/selectors';
 import { wishListFunctionSelector, wishListsByHashSelector } from '../wishlists/selectors';
 import { InventoryWishListRoll } from '../wishlists/wishlists';
 import { FilterContext, SuggestionsContext } from './filter-types';
-import { searchConfigSelector } from './item-search-config';
+import { buildFiltersMap, buildSearchConfig } from './search-config';
 import { makeSearchFilterFactory, parseAndValidateQuery } from './search-filter';
+import advancedFilters from './search-filters/advanced';
+import d1Filters from './search-filters/d1-filters';
+import dupeFilters from './search-filters/dupes';
+import freeformFilters from './search-filters/freeform';
+import itemInfosFilters from './search-filters/item-infos';
+import knownValuesFilters from './search-filters/known-values';
+import loadoutFilters from './search-filters/loadouts';
+import simpleRangeFilters from './search-filters/range-numeric';
+import overloadedRangeFilters from './search-filters/range-overload';
+import simpleFilters from './search-filters/simple';
+import socketFilters from './search-filters/sockets';
+import statFilters from './search-filters/stats';
+import locationFilters from './search-filters/stores';
+import wishlistFilters from './search-filters/wishlist';
+import { suggestionsContextSelector } from './suggestions-generation';
+
+const allFilters = [
+  ...dupeFilters,
+  ...($featureFlags.wishLists ? wishlistFilters : []),
+  ...freeformFilters,
+  ...itemInfosFilters,
+  ...knownValuesFilters,
+  ...d1Filters,
+  ...loadoutFilters,
+  ...simpleRangeFilters,
+  ...overloadedRangeFilters,
+  ...simpleFilters,
+  ...socketFilters,
+  ...statFilters,
+  ...locationFilters,
+  ...advancedFilters,
+];
+
+export const buildItemFiltersMap = memoizeOne((destinyVersion: DestinyVersion) =>
+  buildFiltersMap(destinyVersion, allFilters),
+);
+
+export function buildItemSearchConfig(
+  destinyVersion: DestinyVersion,
+  language: DimLanguage,
+  suggestionsContext: SuggestionsContext = {},
+) {
+  const filtersMap = buildItemFiltersMap(destinyVersion);
+  return buildSearchConfig(language, suggestionsContext, filtersMap);
+}
 
 //
 // Selectors
 //
+
+export const searchConfigSelector = createSelector(
+  destinyVersionSelector,
+  languageSelector,
+  suggestionsContextSelector,
+  buildItemSearchConfig,
+);
 
 /**
  * A selector for the filterContext for a particular destiny version. This must
